Conflicts:
	Application/Manager/CandidaturaManager.php

+++ Application/Service/EliminarPrestadorService.php
@@ -9,18 +9,21 @@ require_once (Conf::getApplicationManagerPath() . 'CandidaturaManager.php');
 require_once (Conf::getApplicationManagerPath() . 'PrestadorManager.php');
 require_once (Conf::getApplicationManagerPath() . 'SessionManager.php');
 $exist = SessionManager::existSession('email');
+$tipo = SessionManager::existSession('tipoUser');
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
 
-if ($exist) {
-    $id = filter_input(INPUT_GET, 'idPrestador');
-    $managerCand = new CandidaturaManager();
-    $managerCand->deleteCandidaturaByIdPrestador($id);
-    $managerPrest = new PrestadorManager();
-    $managerPrest->deletePrestadorById($id);
-    echo 'Eliminado';
+if ($exist && $tipo) {
+    if (SessionManager::getSessionValue('tipoUser') === 'administrador') {
+        $id = filter_input(INPUT_GET, 'idPrestador');
+        $managerCand = new CandidaturaManager();
+        $managerCand->deleteCandidaturaByIdPrestador($id);
+        $managerPrest = new PrestadorManager();
+        $managerPrest->deletePrestadorById($id);
+        echo 'Eliminado';
+    }
 }
